+++ cobs/cortex_file.hpp
@@ -46,7 +46,7 @@ public:
 
     static void check_magic_number(std::istream& is, std::string path) {
         std::string magic_word = "CORTEX";
-        for (size_t i = 0; i < magic_word.size(); i++) {
+        for (uint64_t i = 0; i < magic_word.size(); i++) {
             if (is.get() != magic_word[i]) {
                 throw std::invalid_argument(
                           "CortexFile: magic number not found @ " + path);
@@ -66,18 +66,18 @@ public:
         if (num_colors_ != 1)
             die("Invalid number of colors (" << num_colors_ << "), must be 1");
 
-        for (size_t i = 0; i < num_colors_; i++) {
+        for (uint64_t i = 0; i < num_colors_; i++) {
             uint32_t mean_read_length = cast_advance<uint32_t>(is);
             uint64_t total_length = cast_advance<uint64_t>(is);
             tlx::unused(mean_read_length, total_length);
         }
-        for (size_t i = 0; i < num_colors_; i++) {
+        for (uint64_t i = 0; i < num_colors_; i++) {
             auto document_name_length = cast_advance<uint32_t>(is);
             name_.resize(document_name_length);
             is.read(const_cast<char*>(name_.data()), document_name_length);
         }
         is.ignore(16 * num_colors_);
-        for (size_t i = 0; i < num_colors_; i++) {
+        for (uint64_t i = 0; i < num_colors_; i++) {
             is.ignore(12);
             auto length_graph_name = cast_advance<uint32_t>(is);
             is.ignore(length_graph_name);
@@ -95,18 +95,18 @@ public:
         pos_data_end_ = is.tellg();
     }
 
-    size_t num_kmers() const {
+    uint64_t num_kmers() const {
         return (pos_data_end_ - pos_data_begin_)
                / (8 * num_words_per_kmer_ + 5 * num_colors_);
     }
 
     template <typename Callback>
-    void process_terms(size_t term_size, Callback callback) {
+    void process_terms(uint64_t term_size, Callback callback) {
         std::string kmer(kmer_size_, 0);
-        static const size_t kmer_packed_size = (kmer_size_ + 3) / 4;
+        static const uint64_t kmer_packed_size = (kmer_size_ + 3) / 4;
 
         // bytes per k-mer from 64-bit words per k-mer (<W>)
-        size_t bytes_per_kmer = sizeof(uint64_t) * num_words_per_kmer_;
+        uint64_t bytes_per_kmer = sizeof(uint64_t) * num_words_per_kmer_;
         die_unless(bytes_per_kmer >= kmer_packed_size);
 
         std::vector<uint8_t> kmer_data(bytes_per_kmer);
@@ -114,7 +114,7 @@ public:
         is_.clear();
         is_.seekg(pos_data_begin_);
 
-        size_t r = num_kmers();
+        uint64_t r = num_kmers();
         while (r != 0) {
             --r;
             if (!is_.good())
@@ -127,7 +127,7 @@ public:
 
             // from KMer::to_string()
             kmer.clear();
-            for (size_t i = 0; i < kmer_packed_size; ++i) {
+            for (uint64_t i = 0; i < kmer_packed_size; ++i) {
                 if (TLX_UNLIKELY(i == 0 && kmer_size_ % 4 != 0)) {
                     // fragment of last k-mer
                     kmer += kmer_byte_to_base_pairs[
@@ -139,7 +139,7 @@ public:
                 }
             }
 
-            for (size_t i = 0; i + term_size <= kmer_size_; ++i) {
+            for (uint64_t i = 0; i + term_size <= kmer_size_; ++i) {
                 callback(tlx::string_view(kmer.data() + i, term_size));
             }
         }
