+++ cobs/fastq_file.hpp
@@ -60,7 +60,7 @@ public:
         sequence_count_ = 0;
         size_ = 0;
 
-        size_t line_num = 0;
+        uint64_t line_num = 0;
         while (std::getline(is, line)) {
             size_ += line.size() + 1;
 
@@ -72,7 +72,7 @@ public:
             }
             else if (line_num % 4 == 1) {
                 // sequence/read line
-                size_t sequence_size = line.size();
+                uint64_t sequence_size = line.size();
                 sequence_size_hist_[sequence_size]++;
                 sequence_count_++;
             }
@@ -123,14 +123,14 @@ public:
     bool read_cache_file() {
         std::ifstream is(cache_path());
         if (!is.good()) return false;
-        size_t hist_size;
+        uint64_t hist_size;
         stream_get_pod(is, size_);
         stream_get_pod(is, sequence_count_);
         stream_get_pod(is, hist_size);
         LOG1 << "FastqFile: loading index " << cache_path()
              << " [" << sequence_count_ << " subsequences]";
-        for (size_t i = 0; i < hist_size; ++i) {
-            size_t size, count;
+        for (uint64_t i = 0; i < hist_size; ++i) {
+            uint64_t size, count;
             stream_get_pod(is, size);
             stream_get_pod(is, count);
             sequence_size_hist_[size] = count;
@@ -139,13 +139,13 @@ public:
     }
 
     //! return estimated size of a fastq document
-    size_t size() {
+    uint64_t size() {
         return size_;
     }
 
     //! return number of q-grams in document
-    size_t num_terms(size_t q) {
-        size_t total = 0;
+    uint64_t num_terms(uint64_t q) {
+        uint64_t total = 0;
         for (const auto& p : sequence_size_hist_) {
             total += p.second * (p.first < q ? 0 : p.first - q + 1);
         }
@@ -153,10 +153,10 @@ public:
     }
 
     template <typename Callback>
-    void process_terms(std::istream& is, size_t term_size, Callback callback) {
+    void process_terms(std::istream& is, uint64_t term_size, Callback callback) {
         std::string line;
 
-        size_t line_num = 0;
+        uint64_t line_num = 0;
         while (std::getline(is, line)) {
             if (line_num % 4 == 0) {
                 if (line.size() == 0 || line[0] != '@') {
@@ -166,7 +166,7 @@ public:
             }
             else if (line_num % 4 == 1) {
                 // process terms in sequence/read line.
-                for (size_t i = 0; i + term_size <= line.size(); ++i) {
+                for (uint64_t i = 0; i + term_size <= line.size(); ++i) {
                     callback(tlx::string_view(line.data() + i, term_size));
                 }
             }
@@ -184,7 +184,7 @@ public:
     }
 
     template <typename Callback>
-    void process_terms(size_t term_size, Callback callback) {
+    void process_terms(uint64_t term_size, Callback callback) {
         is_.clear();
         is_.seekg(0);
         die_unless(is_.good());
@@ -204,11 +204,11 @@ private:
     //! path
     std::string path_;
     //! size in bytes
-    size_t size_;
+    uint64_t size_;
     //! number of sub-sequences
-    size_t sequence_count_;
+    uint64_t sequence_count_;
     //! histogram of sub-sequence sizes
-    std::map<size_t, size_t> sequence_size_hist_;
+    std::map<uint64_t, uint64_t> sequence_size_hist_;
 };
 
 } // namespace cobs
