+++ cobs/query/index_file.hpp
@@ -22,8 +22,8 @@ public:
     StreamPos stream_pos_;
 
     virtual void read_from_disk(
-        const std::vector<size_t>& hashes, uint8_t* rows,
-        size_t begin, size_t size, size_t buffer_size) = 0;
+        const std::vector<uint64_t>& hashes, uint8_t* rows,
+        uint64_t begin, uint64_t size, uint64_t buffer_size) = 0;
 
     virtual uint32_t term_size() const = 0;
     virtual uint8_t canonicalize() const = 0;
